ethical_drugs/requirements.txt
# Core dependencies
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings

# Database
SQLAlchemy
psycopg2-binary
alembic
